+++ genno/compat/pyam/computations.py
@@ -1,4 +1,5 @@
 import logging
+from pathlib import Path
 from typing import Callable, Collection, Optional, Union
 
 import pyam
@@ -90,20 +91,22 @@ def concat(*args, **kwargs):
         return genno.computations.concat(*args, **kwargs)
 
 
-def write_report(quantity, path):
-    """Write the report identified by *key* to the file at *path*.
+def write_report(obj, path: Union[str, Path]) -> None:
+    """Write  obj` to the file at `path`.
 
-    If *quantity* is a :class:`pyam.IamDataFrame` and *path* ends with '.csv' or
-    '.xlsx', use :mod:`pyam` methods to write the file to CSV or Excel format,
-    respectively. Otherwise, equivalent to :func:`genno.computations.write_report`.
+    If `obj` is a :class:`pyam.IamDataFrame` and `path` ends with ".csv" or ".xlsx",
+    use :mod:`pyam` methods to write the file to CSV or Excel format, respectively.
+    Otherwise, equivalent to :func:`genno.computations.write_report`.
     """
-    if not isinstance(quantity, pyam.IamDataFrame):
-        return genno.computations.write_report(quantity, path)
+    if not isinstance(obj, pyam.IamDataFrame):
+        return genno.computations.write_report(obj, path)
+
+    path = Path(path)
 
     if path.suffix == ".csv":
-        quantity.to_csv(path)
+        obj.to_csv(path)
     elif path.suffix == ".xlsx":
-        quantity.to_excel(path, merge_cells=False)
+        obj.to_excel(path, merge_cells=False)
     else:
         raise ValueError(
             f"pyam.IamDataFrame can be written to .csv or .xlsx, not {path.suffix}"