+++ genno/tests/test_core.py
@@ -1,3 +1,5 @@
+import re
+
 from genno import configure
 from genno.util import REPLACE_UNITS
 
@@ -5,7 +7,7 @@ from genno.util import REPLACE_UNITS
 def test_configure_units(caplog):
     # Warning is logged on invalid definitions
     configure(units=dict(define="0 = [0] * %"))
-    assert 'missing unary operator "*"' in caplog.messages
+    assert re.match(r'missing unary operator "."', caplog.messages[0])
 
     # Unit replacements are stored
     configure(units=dict(replace={"foo": "bar"}))